Codebase simplified + BOX ADDED
Instead of switching the whole scene between dimensions, the game will now simply activate or deactivate the objects to give the illusion of a different room. This was done to simplify the process of making literally anything. Also, due to this change, the Box was finally added alongside a simple inventory mechanic which only supports 1 item for now.